Feature/double precision

+++ src/canari/data_process.py
@@ -94,7 +94,7 @@ class DataProcess:
         self.time_covariates = time_covariates
         self.output_col = output_col
 
-        data = data.astype("float32")
+        # data = data.astype("float32")
         self.data = data.copy()
         self.scale_const_mean, self.scale_const_std = None, None
 
@@ -119,7 +119,7 @@ class DataProcess:
             "hour_of_day": lambda idx: idx.hour,
             "day_of_week": lambda idx: idx.dayofweek,
             "day_of_year": lambda idx: idx.dayofyear,
-            "week_of_year": lambda idx: idx.isocalendar().week,
+            "week_of_year": lambda idx: idx.isocalendar().week.astype(float),
             "month_of_year": lambda idx: idx.month,
             "quarter_of_year": lambda idx: idx.quarter,
         }
@@ -129,7 +129,7 @@ class DataProcess:
 
         for cov in dict.fromkeys(self.time_covariates):
             vals = allowed[cov](self.data.index)
-            self.data[cov] = np.array(vals, dtype=np.float32)
+            self.data[cov] = np.array(vals)
 
     def _get_split_start_end_indices(self):
         """
